requirements.txt
streamlit>=1.33.1
pdfplumber>=0.10.0
python-dotenv>=1.0.0
openai>=0.30.0
pydantic>=1.10.0

